cognee/requirements.txt
cognee
litellm
pdfplumber
python-dotenv
greenlet==3.1.1
